\name{gonot3}

\alias{gonot3}

\title{
Calculate biting rate for Anopheles gambiae s.l.
}

\description{
Calculate biting rate for Anopheles gambiae s.l. based on the number of days it takes to complete a gonothropic cycle
}

\usage{
gonot3(T)
}

\arguments{
  \item{T}{Air temperature in C}
}

\details{

\code{gonot3} calculate the biting rate for Anopheles gambiae s.l.
}

\value{
A vector with biting rate per day
}

\references{
}

\author{
Torleif Markussen Lunde (Centre for International Health/Geophysical Institute, University of Bergen, Norway)
}

\note{}

\seealso{
}

\examples{
bite <- gonot3(20)
}


\keyword{malaria}
